+++ package.json
@@ -14,11 +14,6 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
-    "./browser": {
-      "types": "./dist/index.browser.d.ts",
-      "import": "./dist/index.browser.js",
-      "default": "./dist/index.browser.js"
-    },
     "./async": {
       "types": "./dist/async/index.d.ts",
       "import": "./dist/async/index.js",
